+++ CHANGELOG.md
@@ -1,5 +1,94 @@
 # Changelog
 
+## 1.8.0 (2020-08-10)
+
+### Added
+
+- In `iServiceData` there is a new flag (`isComponent`), so behaviour can change if the extension is working with a component vs. a service. For future use.
+- New internal constant (`CAN_CANCEL_OPERATIONS`) to configure if operations that are running can be cancelled. It's value is false.
+
+### Changed
+
+- Improve all progress notifications with progress bars.
+- Bundle extension to improve size and loading time.
+- Update packages to avoid security flaws.
+
+## 1.7.2 (2020-06-26)
+
+### Fixed and changed
+
+- DeprecateService:
+  - Version in finish message has been corrected from `MajorVersion.MajorVersion` to `MajorVersion.MinorVersion`.
+- GAMaker:
+  - Major version and minor version now appear correctly in release note title.
+  - Text has been corrected to avoid unnecessary uppercase letters in release notes.
+  - Version in finish message has been corrected from `MajorVersion.MajorVersion` to `MajorVersion.MinorVersion`.
+- NewTask:
+  - New tasks by default now are created with the preview tag set to true.
+- RestorePackages:
+  - Version in finish message has been corrected from `MajorVersion.MajorVersion` to `MajorVersion.MinorVersion`.
+- SwaggerGenerator:
+  - Service version now is correct.
+  - Link to Features Team wiki has been updated so the new one is in use.
+  - Secret section generation was not working properly due to different names of the section that can be found after secret section, which is use cases.
+  - A whitespace now is correctly placed between sentences in the parameter's description.
+
+## 1.7.1 (2020-05-11)
+
+### Changed
+
+- Modify regular expressions so the functionality can be used from `Components` folder too.
+- Update finish messages to unify them with complete version number.
+- Context menu image now has only default + extension options (without extra items).
+
+## 1.7.0 (2020-05-06)
+
+### Added
+
+- New functionality: create a new service (initialize a service folder).
+- Added a new function `ValidateFolderPath` for the folder containing all services.
+- New setting:
+  - Regular expression to validate service version number.
+- New functionality in common:
+  - Function to replace all instances of a string. Unlike `string.replace` that only changes the first occurrence.
+  - Function to read all files and directories in a given path. It returns an object with two lists, files and directories.
+
+### Changed
+
+- Last changelog entry was marked as "Added" when it must be "Changed", fixed.
+- The Regular Expressions that matches directories and tasks now supports `CloudFoundation` as origin path.
+- Changed the name of the Regular Expression to validate service folder from `FolderRegExp` to `FolderServiceRegExp`.
+- Rename `ValidateFolderPath` to `ValidateServiceFolderPath`, as it's more accurate.
+
+### Fixed
+
+- The Regular Expressions now detect better wrong path structures.
+
+## 1.6.2 (2020-04-27)
+
+### Changed
+
+- Swagger Generator: Integer parameters now have the correct type.
+
+## 1.6.1 (2020-04-17)
+
+### Added
+
+- Markdownlint configuration file to remove CHANGELOG alerts:
+  - MD024: duplicate headers.
+  - MD013: line length.
+- Status information in all services.
+
+### Changed
+
+- Images now working on README file after setting repository on packages.json.
+- Change progress bar (not working properly) for custom status bar text.
+- Improved features section in README.
+
+### Removed
+
+- Unused common method (Succeeded).
+
 ## 1.6.0 (2020-04-05)
 
 ### Added
@@ -62,7 +151,7 @@
 ### Changed
 
 - Created internal function `ValidatePath` to make it more reusable.
-- Added two new funtions `ValidateFolderPath` and `ValidateTaskPath` that, internally, use the old `ValidatePath`.
+- Added two new functions `ValidateFolderPath` and `ValidateTaskPath` that, internally, use the old `ValidatePath`.
 - SwaggerGenerator: the summary of path section now comes from the description of the task.json file of the operation instead, of the friendly name.
 - README to reflect new operation and configurations.